app/requirements.txt
streamlit==1.4.2
google-generativeai==0.5.0
python-dotenv==1.0.1
vertexai
langchain==0.3.4
langchain-community==0.3.3
beautifulsoup4==4.12.3
trafilatura==1.12.2
faiss-cpu==1.9.0
langchain-openai==0.2.5